requirements.txt
discord.py
saucenao-api
configparse
requests
praw
asyncpraw
psycopg2
mcstatus
tweepy
mega.py
discord-components
